backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
redis==5.0.1
openai==1.3.0
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.0
Pillow==10.0.1 